Revert "Merge branch 'main' into Create_ImGuizmo"
This reverts commit 1ac15e93ab25c0e8255471df74f3af535c7a63b2, reversing
changes made to 4d75992c3eb2e4022c4b95ec408f78384dee99f0.

+++ Project/Engine/CTransform.cpp
@@ -30,7 +30,7 @@ void CTransform::finaltick()
 	float y = m_vRelativeScale.y ;
 	float z = m_vRelativeScale.z ;
 
-	m_fBoundingRadius = (float)sqrt(pow(x, 2) + pow(y, 2) + pow(z, 2));
+	m_fBoundingRadius = sqrt(pow(x, 2) + pow(y, 2) + pow(z, 2));
 
 	m_matWorldBoundingScale = XMMatrixIdentity();
 	m_matWorldBoundingScale = XMMatrixScaling(m_fBoundingRadius, m_fBoundingRadius, m_fBoundingRadius);
